requirements.txt
fastapi~=0.115.4
uvicorn
sqlalchemy~=2.0.36
asyncpg
python-dotenv~=1.0.1
pydantic~=2.9.2
passlib[bcrypt]~=1.7.4
websockets~=11.0.3
httpx~=0.24.1
python-jose~=3.3.0
openai~=0.27.8